requirements.txt
selenium
pyppeteer
websockets==6.0
requests
pillow